refactor(webhook): split do god-file em módulos — index vira roteador (Onda 6b)
telegram-webhook/index.ts tinha 2.427 linhas acumulando TUDO (extração LLM,
callbacks, item 15, onboarding, comandos, vínculo). Cada feature nova era um
branch no mesmo serve(), e merges colidiam (item 15 × onboarding). Split
MECÂNICO (código movido byte a byte por script; só headers de import e
footers de export gerados; validador de referências órfãs: 8/8 OK):

- config.ts (env/constantes) · types.ts (interfaces)
- telegram-api.ts (chamadas + mensagens padrão do bot)
- picks.ts (item 15: unidade, registro, recibo, prompt de valor)
- callbacks.ts (TODOS os botões: settle/mute/mutew/fix/regbet/stk/prefs)
- extraction.ts (pipeline print/texto/áudio → aposta + limites diários)
- users.ts (vínculo Telegram↔conta)
- index.ts: 2.427 → 537 linhas (roteador + fluxo de mensagem: link_,
  comandos, interceptador de valor)

CI ganha `deno check` do grafo do webhook (pega import/ref quebrada entre
módulos antes do deploy) — junto dos 77 testes existentes.

Zero mudança de comportamento. Gate: deno check + deploy bundling + ensaio
manual no staging (botões, /comandos, registro por texto).
