+++ supabase/functions/telegram-webhook/index.ts
@@ -1,1923 +1,31 @@
+// ============================================================
+// telegram-webhook — roteador de updates do bot (Onda 6b: modularizado)
+// ============================================================
+// index.ts é o ROTEADOR: auth do webhook, dedupe de update_id e o fluxo de
+// mensagem (onboarding link_, /comandos, interceptador de valor, registro de
+// aposta). O resto vive em módulos:
+//   config.ts (env/constantes) · types.ts · telegram-api.ts (chamadas/copy)
+//   callbacks.ts (todos os botões) · picks.ts (item 15) · extraction.ts
+//   (pipeline LLM + limites) · users.ts (vínculo) · stake.ts/prefs.ts (puros)
 import { serve } from "https://deno.land/std@0.168.0/http/server.ts"
 import { createClient } from "npm:@supabase/supabase-js@2"
-import { normalizePhoneNumber, normalizePhoneCandidates, maskPhone } from "../shared/phone.ts"
-import { generateTraceId, identifyUser, trackEvent, trackLLMGeneration } from "../shared/posthog.ts"
+import { maskPhone } from "../shared/phone.ts"
+import { generateTraceId, identifyUser, trackEvent } from "../shared/posthog.ts"
 import { isBareNumber, parseStake } from "./stake.ts"
 import { prefsKeyboard, prefsText } from "./prefs.ts"
+import { TELEGRAM_BOT_TOKEN, TELEGRAM_WEBHOOK_SECRET } from "./config.ts"
+import type { TelegramMessage, TelegramCallbackQuery } from "./types.ts"
+import {
+  sendTelegramMessage, getTelegramFileUrl, sendContactRequest, sendWelcomeMessageTelegram,
+} from "./telegram-api.ts"
+import { registerPickBet, pickReceiptHtml } from "./picks.ts"
+import { handleCallbackQuery } from "./callbacks.ts"
+import { transcribeAudio, processImage, processMessage } from "./extraction.ts"
+import { findUserByTelegram, findUserByPhone } from "./users.ts"
 
-const OPENAI_API_KEY = Deno.env.get("OPENAI_API_KEY")
-const OPENAI_API_URL = "https://api.openai.com/v1"
-const TELEGRAM_BOT_TOKEN = Deno.env.get("TELEGRAM_BOT_TOKEN")
-const TELEGRAM_WEBHOOK_SECRET = Deno.env.get("TELEGRAM_WEBHOOK_SECRET")
-
-const TELEGRAM_API_BASE = TELEGRAM_BOT_TOKEN
-  ? `https://api.telegram.org/bot${TELEGRAM_BOT_TOKEN}`
-  : ""
-
-const DAILY_BET_LIMIT = 3
-const TIMEZONE_GMT3 = "America/Cuiaba"
-const BETS_DASHBOARD_URL = "https://www.smartbetting.app/bets"
 
 // Simple in-memory guard to avoid reprocessing the same update_id on retries
 const processedUpdateIds = new Set<number>()
-
-interface TelegramFile {
-  file_id: string
-  file_unique_id: string
-  file_size?: number
-  file_path?: string
-}
-
-interface TelegramUser {
-  id: number
-  is_bot: boolean
-  first_name?: string
-  last_name?: string
-  username?: string
-  language_code?: string
-  phone_number?: string
-}
-
-interface TelegramMessage {
-  message_id: number
-  date: number
-  chat: { id: number; type: string; title?: string; username?: string; first_name?: string; last_name?: string }
-  from?: TelegramUser
-  text?: string
-  caption?: string
-  contact?: { phone_number: string; user_id?: number }
-  photo?: TelegramFile[]
-  voice?: TelegramFile
-  audio?: TelegramFile
-  document?: TelegramFile
-  // Telegram sends thumbnails as "thumbnail" in documents; not used directly
-  thumbnail?: TelegramFile
-  // resposta a uma mensagem do bot (force_reply do "Outro valor")
-  reply_to_message?: { message_id: number }
-}
-
-// Toque num botão inline (lembrete de liquidação do notify-settlement)
-interface TelegramCallbackQuery {
-  id: string
-  from?: TelegramUser
-  message?: { message_id: number; chat: { id: number } }
-  data?: string
-}
-
-interface ProcessedBet {
-  bet_type: string
-  sport: string
-  league?: string
-  betting_market?: string
-  matches: Array<{
-    description: string
-    bet_description: string
-    odds: number
-    match_date?: string
-    is_combined_odd?: boolean
-  }>
-  stake_amount: number
-  bet_date: string
-  odds_are_individual?: boolean
-}
-
-// JSON Schema for structured outputs (parity with WhatsApp)
-const BETTING_INFO_SCHEMA = {
-  type: "object",
-  properties: {
-    bet_type: { type: "string", enum: ["single", "multiple", "system"] },
-    sport: { type: "string" },
-    league: { type: ["string", "null"] },
-    matches: {
-      type: "array",
-      minItems: 1,
-      items: {
-        type: "object",
-        required: ["description", "bet_description", "odds", "match_date", "is_combined_odd"],
-        properties: {
-          description: { type: "string" },
-          bet_description: { type: "string" },
-          odds: { type: "number", minimum: 1.01 },
-          match_date: { type: ["string", "null"] },
-          is_combined_odd: { type: "boolean" }
-        },
-        additionalProperties: false
-      }
-    },
-    stake_amount: { type: "number", minimum: 0 },
-    bet_date: { type: "string" },
-    odds_are_individual: { type: "boolean" },
-    betting_market: { type: "string" }
-  },
-  required: ["bet_type", "sport", "matches", "stake_amount", "bet_date", "odds_are_individual", "league", "betting_market"],
-  additionalProperties: false
-}
-
-// Calculator tool definition (parity with WhatsApp)
-const CALCULATE_ODDS_TOOL = {
-  type: "function",
-  function: {
-    name: "calculate_multiple_odds",
-    description: "Calculate the combined odds for multiple bets by multiplying individual odds together. Use this tool when you need to verify or calculate odds for a multiple bet, especially when uncertain about whether odds are individual or already combined.",
-    parameters: {
-      type: "object",
-      properties: {
-        odds: { type: "array", items: { type: "number" }, description: "Array of individual odds to multiply together" }
-      },
-      required: ["odds"]
-    }
-  }
-}
-
-// Calculator function implementation
-async function calculateMultipleOdds(odds: number[]): Promise<number> {
-  if (!odds || odds.length === 0) return 1
-  const result = odds.reduce((acc, odd) => {
-    if (odd < 1.01) return acc * 1.01
-    return acc * odd
-  }, 1)
-  return Math.round(result * 100) / 100
-}
-
-async function telegramCall<T = any>(method: string, payload: Record<string, unknown>): Promise<T | null> {
-  if (!TELEGRAM_API_BASE) {
-    console.error("Missing TELEGRAM_BOT_TOKEN")
-    return null
-  }
-
-  const response = await fetch(`${TELEGRAM_API_BASE}/${method}`, {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(payload)
-  })
-
-  if (!response.ok) {
-    const errorText = await response.text()
-    console.error(`Telegram API error ${method}: ${response.status} - ${errorText}`)
-    return null
-  }
-
-  return await response.json()
-}
-
-async function sendTelegramMessage(chatId: string | number, text: string, options: Record<string, unknown> = {}): Promise<void> {
-  await telegramCall("sendMessage", {
-    chat_id: chatId,
-    text,
-    parse_mode: "Markdown",
-    ...options
-  })
-}
-
-async function sendTelegramPhoto(chatId: string | number, photoUrl: string, caption?: string): Promise<void> {
-  await telegramCall("sendPhoto", {
-    chat_id: chatId,
-    photo: photoUrl,
-    caption,
-    parse_mode: "Markdown"
-  })
-}
-
-async function getTelegramFileUrl(fileId: string): Promise<string | null> {
-  if (!TELEGRAM_API_BASE) return null
-  const getFile = await telegramCall<{ result?: TelegramFile }>("getFile", { file_id: fileId })
-  const filePath = getFile?.result?.file_path
-  if (!filePath) return null
-  return `https://api.telegram.org/file/bot${TELEGRAM_BOT_TOKEN}/${filePath}`
-}
-
-async function sendContactRequest(chatId: string | number): Promise<void> {
-  const keyboard = {
-    keyboard: [[{ text: "📲 Enviar meu número", request_contact: true }]],
-    resize_keyboard: true,
-    one_time_keyboard: true
-  }
-
-  const text = [
-    "Para continuar, preciso confirmar seu telefone e vincular sua conta.",
-    "Toque em *Enviar meu número* abaixo para compartilhar seu contato."
-  ].join("\n")
-
-  await sendTelegramMessage(chatId, text, { reply_markup: keyboard })
-}
-
-// Welcome em 2 mensagens (onboarding redesign): (1) valor em 3 linhas,
-// (2) pedido explícito da 1ª aposta com exemplo copiável — o momento-aha
-// é registrar a primeira aposta em <5 min do cadastro.
-async function sendWelcomeMessageTelegram(chatId: string | number, userName?: string): Promise<void> {
-  const welcome = [
-    `✅ *Conectado${userName ? `, ${userName}` : ""}!* Eu sou o Betinho.`,
-    "",
-    "A partir de agora, por aqui você:",
-    "📸 registra apostas por print, texto ou áudio",
-    "📊 acompanha seu ROI real, liquidação e banca",
-    "📬 recebe as oportunidades do dia e avisos"
-  ].join("\n")
-
-  const firstBetAsk = [
-    "*Bora registrar sua primeira aposta?*",
-    "",
-    "Manda um print da aposta (1 por print — se faltar o valor, escreve na mesma mensagem), ou copia e adapta:",
-    "",
-    "`Flamengo x Palmeiras - Flamengo vence - Odd 1.85 - R$ 50`"
-  ].join("\n")
-
-  await sendTelegramMessage(chatId, welcome)
-  await sendTelegramMessage(chatId, firstBetAsk)
-}
-
-async function sendHelpMessageTelegram(chatId: string | number): Promise<void> {
-  const helpMessage = [
-    "🏀 *COMO ENVIAR SUAS APOSTAS:*",
-    "",
-    "*📸 Screenshot da aposta:*",
-    "• 1 aposta por print",
-    "• Se faltar algo (valor), escreva junto",
-    "• Exemplo: [IMAGEM] + \"100 reais\"",
-    "",
-    "*✍️ Texto:*",
-    "`Lakers vs Warriors - LeBron 25+ pontos - Odd 1.85 - R$ 50`",
-    "",
-    "⚠️ *IMPORTANTE:*",
-    "• 1 mensagem = 1 aposta",
-    "• Envie tudo junto (imagem + texto)",
-    "• Se não recebeu confirmação, reenvie a imagem"
-  ].join("\n")
-
-  await sendTelegramMessage(chatId, helpMessage)
-}
-
-async function sendPaywallMessageTelegram(chatId: string | number): Promise<void> {
-  const paywallUrl = "https://smartbetting.app/paywall"
-  const paywallMessage = [
-    "🚫 *Limite Diário Atingido!*",
-    "",
-    `Você atingiu o limite de ${DAILY_BET_LIMIT} apostas grátis por dia.`,
-    "",
-    "Para continuar apostando, acesse:",
-    paywallUrl
-  ].join("\n")
-
-  await sendTelegramMessage(chatId, paywallMessage)
-}
-
-async function sendConfirmationMessageTelegram(
-  chatId: string | number,
-  betDetails: {
-    bet_type: string
-    sport: string
-    league?: string | null
-    match_description: string
-    bet_description: string
-    odds: number
-    stake_amount: number
-    potential_return: number
-  }
-): Promise<void> {
-  const betTypeText: Record<string, string> = {
-    single: "Simples",
-    multiple: "Múltipla",
-    system: "Sistema"
-  }
-
-  const dashboardUrl = "https://www.smartbetting.app/bets"
-
-  const confirmationMessage = [
-    "🎯 *Aposta Registrada com Sucesso!*",
-    "",
-    "📊 *Detalhes da Aposta:*",
-    `• *Tipo:* ${betTypeText[betDetails.bet_type] || betDetails.bet_type}`,
-    `• *Esporte:* ${betDetails.sport}${betDetails.league ? ` (${betDetails.league})` : ""}`,
-    `• *Jogo:* ${betDetails.match_description}`,
-    `• *Aposta:* ${betDetails.bet_description}`,
-    `• *Odds:* ${betDetails.odds}`,
-    `• *Valor:* R$ ${betDetails.stake_amount.toFixed(2)}`,
-    `• *Retorno Potencial:* R$ ${betDetails.potential_return.toFixed(2)}`,
-    "",
-    "✅ Sua aposta foi salva no dashboard e você pode acompanhar o resultado em tempo real!",
-    "",
-    `🔗 Acesse seu dashboard: ${dashboardUrl}`,
-    "",
-    "⚠️ *IMPORTANTE:*",
-    "Se você enviou uma imagem de aposta e não recebeu esta mensagem, envie a imagem novamente."
-  ].join("\n")
-
-  await sendTelegramMessage(chatId, confirmationMessage)
-}
-
-// ============================================================
-// Liquidação por botão — [✅ Green] [❌ Red] do lembrete (notify-settlement)
-// ============================================================
-
-// Escapa HTML (os lembretes e suas edições usam parse_mode HTML)
-function escHtml(s: unknown): string {
-  return String(s ?? "").replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;")
-}
-
-const formatBRL = (v: number) => `R$ ${Number(v ?? 0).toFixed(2).replace(".", ",")}`
-
-async function answerCallbackQuery(callbackQueryId: string, text?: string): Promise<void> {
-  await telegramCall("answerCallbackQuery", { callback_query_id: callbackQueryId, text })
-}
-
-// Reescreve a mensagem do lembrete como recibo da liquidação (some o teclado)
-async function editSettledMessage(chatId: string | number, messageId: number, html: string): Promise<void> {
-  await telegramCall("editMessageText", {
-    chat_id: chatId,
-    message_id: messageId,
-    text: html,
-    parse_mode: "HTML",
-    disable_web_page_preview: true
-  })
-}
-
-function settledHtml(bet: any, status: string): string {
-  const jogo = bet.match_description ? `\n${escHtml(bet.match_description)}` : ""
-  const base = `<b>${escHtml(bet.bet_description)}</b>${jogo}`
-  if (status === "won") {
-    const lucro = (bet.potential_return ?? 0) - (bet.stake_amount ?? 0)
-    return `✅ <b>Green registrado!</b>\n\n${base}\nLucro: +${formatBRL(lucro)}\n\nBanca atualizada 📊 ${BETS_DASHBOARD_URL}`
-  }
-  if (status === "lost") {
-    return `❌ <b>Red registrado.</b>\n\n${base}\nPrejuízo: −${formatBRL(bet.stake_amount)}\n\nFaz parte. Banca atualizada 📊 ${BETS_DASHBOARD_URL}`
-  }
-  return `Registrada como <b>${escHtml(status)}</b>.\n\n${base}`
-}
-
-// ============================================================
-// Registrar aposta da "oportunidade do dia" (item 15) — botões de stake
-// ============================================================
-
-// Unidade efetiva do usuário em R$ (fixo = valor; percentual = % da banca).
-// null = não configurada → só oferece "Outro valor".
-async function effectiveUnit(supabase: any, userId: string): Promise<number | null> {
-  const { data } = await supabase
-    .from("users")
-    .select("unit_value, unit_calculation_method, bank_amount")
-    .eq("id", userId)
-    .maybeSingle()
-  const u = Number(data?.unit_value)
-  if (!u || u <= 0) return null
-  if (data.unit_calculation_method === "percentual") {
-    const bank = Number(data.bank_amount)
-    if (!bank || bank <= 0) return null
-    return Math.round(bank * (u / 100) * 100) / 100
-  }
-  return u
-}
-
-// Cria a aposta pendente a partir de um pick do daily. channel 'futebol' marca
-// a origem (loop Análise→Betinho); processed_data guarda o pick pra auditoria.
-async function registerPickBet(supabase: any, userId: string, pick: any, stake: number) {
-  const odds = Number(pick.odds)
-  return await supabase
-    .from("bets")
-    .insert({
-      user_id: userId,
-      bet_type: "single",
-      sport: pick.sport || "Futebol",
-      league: pick.league || null,
-      betting_market: pick.betting_market || null,
-      match_description: pick.match_description,
-      bet_description: pick.bet_description,
-      odds,
-      stake_amount: stake,
-      potential_return: Math.round(stake * odds * 100) / 100,
-      status: "pending",
-      bet_date: new Date().toISOString(),
-      match_date: pick.match_date || null,
-      channel: "futebol",
-      raw_input: "[oportunidade do dia]",
-      processed_data: { source: "daily_opportunity", pick_id: pick.id }
-    })
-    .select("id")
-    .maybeSingle()
-}
-
-function pickReceiptHtml(pick: any, stake: number): string {
-  const odds = Number(pick.odds)
-  return [
-    `✅ <b>Registrado no Betinho!</b>`,
-    "",
-    `<b>${escHtml(pick.bet_description)}</b>`,
-    `${escHtml(pick.match_description)}`,
-    `${formatBRL(stake)} · odd ${odds} · retorno ${formatBRL(Math.round(stake * odds * 100) / 100)}`,
-    "",
-    `Tá pendente — quando o jogo acabar, eu fecho pra você. 📊 ${BETS_DASHBOARD_URL}`
-  ].join("\n")
-}
-
-// pergunta o valor por force_reply e guarda o prompt (o interceptador lê depois)
-async function sendStakePrompt(supabase: any, chatId: string | number, userId: string, pickId: string, text: string): Promise<void> {
-  const sent = await telegramCall<{ result?: { message_id: number } }>("sendMessage", {
-    chat_id: chatId,
-    text,
-    parse_mode: "HTML",
-    reply_markup: { force_reply: true, input_field_placeholder: "Ex: 50" }
-  })
-  const promptId = sent?.result?.message_id
-  if (promptId) {
-    await supabase.from("stake_prompts").insert({
-      chat_id: String(chatId), message_id: promptId, pick_id: pickId, user_id: userId
-    })
-  }
-}
-
-// parseStake/isBareNumber vivem em ./stake.ts (Onda 3 — testados no CI)
-
-async function handleCallbackQuery(
-  supabase: any,
-  cq: TelegramCallbackQuery,
-  traceId: string
-): Promise<Response> {
-  const ok = (msg: string) =>
-    new Response(JSON.stringify({ success: true, message: msg }), {
-      headers: { "Content-Type": "application/json" },
-      status: 200
-    })
-
-  const data = cq.data || ""
-  const chatId = cq.message?.chat?.id
-  const messageId = cq.message?.message_id
-
-  if (!chatId || !messageId) {
-    await answerCallbackQuery(cq.id).catch(() => {})
-    return ok("callback without message")
-  }
-
-  const user = await findUserByTelegram(
-    supabase,
-    cq.from?.id ? String(cq.from.id) : undefined,
-    String(chatId),
-    traceId
-  )
-  if (!user) {
-    await answerCallbackQuery(cq.id, "Não achei sua conta. Manda /start pra sincronizar.")
-    return ok("callback user not found")
-  }
-
-  // 🔕 mute:<bet_id> — para os lembretes; mantém os botões de liquidar desta aposta
-  if (data.startsWith("mute:")) {
-    const betId = data.slice("mute:".length)
-    await supabase.from("users").update({ settlement_reminders_muted: true }).eq("id", user.id)
-    await telegramCall("editMessageReplyMarkup", {
-      chat_id: chatId,
-      message_id: messageId,
-      reply_markup: {
-        inline_keyboard: [
-          [
-            { text: "✅ Green", callback_data: `settle:${betId}:won` },
-            { text: "❌ Red", callback_data: `settle:${betId}:lost` }
-          ],
-          [{ text: "Outras opções ↗", url: BETS_DASHBOARD_URL }]
-        ]
-      }
-    })
-    await answerCallbackQuery(cq.id, "🔕 Ok, sem mais lembretes. Reative mandando /lembretes.")
-    await trackEvent(
-      "settlement_reminders_muted",
-      { via: "button", channel: "telegram" },
-      user.id,
-      traceId
-    ).catch(() => {})
-    return ok("reminders muted")
-  }
-
-  // mutew — botão "Silenciar resumo" da DM do resumo semanal (item 04).
-  // Opt-out independente da liquidação; /resumo reativa.
-  if (data === "mutew") {
-    await supabase.from("users").update({ weekly_summary_muted: true }).eq("id", user.id)
-    // remove só a linha do silenciar; "Ver minha banca" continua útil
-    await telegramCall("editMessageReplyMarkup", {
-      chat_id: chatId,
-      message_id: messageId,
-      reply_markup: {
-        inline_keyboard: [[{ text: "Ver minha banca", url: BETS_DASHBOARD_URL }]]
-      }
-    })
-    await answerCallbackQuery(cq.id, "Ok, sem mais resumos semanais. Pra voltar: /resumo")
-    await trackEvent(
-      "weekly_summary_muted",
-      { via: "button", channel: "telegram" },
-      user.id,
-      traceId
-    ).catch(() => {})
-    return ok("weekly summary muted")
-  }
-
-  // prefliq / prefres — toggles do centro de controle /mensagens (Onda 6).
-  // Alterna a preferência e REESCREVE a própria mensagem com o estado novo.
-  if (data === "prefliq" || data === "prefres") {
-    const { data: cur } = await supabase
-      .from("users").select("settlement_reminders_muted, weekly_summary_muted").eq("id", user.id).maybeSingle()
-    const p = {
-      settlementMuted: cur?.settlement_reminders_muted ?? false,
-      weeklyMuted: cur?.weekly_summary_muted ?? false,
-    }
-    let toast: string
-    if (data === "prefliq") {
-      p.settlementMuted = !p.settlementMuted
-      await supabase.from("users").update({ settlement_reminders_muted: p.settlementMuted }).eq("id", user.id)
-      toast = p.settlementMuted ? "🔕 Liquidação silenciada." : "🔔 Liquidação reativada."
-      await trackEvent(
-        p.settlementMuted ? "settlement_reminders_muted" : "settlement_reminders_unmuted",
-        { via: "prefs", channel: "telegram" }, user.id, traceId
-      ).catch(() => {})
-    } else {
-      p.weeklyMuted = !p.weeklyMuted
-      await supabase.from("users").update({ weekly_summary_muted: p.weeklyMuted }).eq("id", user.id)
-      toast = p.weeklyMuted ? "Resumo semanal silenciado." : "📊 Resumo semanal reativado."
-      await trackEvent(
-        p.weeklyMuted ? "weekly_summary_muted" : "weekly_summary_unmuted",
-        { via: "prefs", channel: "telegram" }, user.id, traceId
-      ).catch(() => {})
-    }
-    await telegramCall("editMessageText", {
-      chat_id: chatId,
-      message_id: messageId,
-      text: prefsText(p),
-      parse_mode: "HTML",
-      disable_web_page_preview: true,
-      reply_markup: { inline_keyboard: prefsKeyboard(p) },
-    })
-    await answerCallbackQuery(cq.id, toast)
-    return ok("prefs toggled")
-  }
-
-  // 📋 regbet:<pick_id> — tocou "Registrar no Betinho" no daily
-  if (data.startsWith("regbet:")) {
-    const pickId = data.slice("regbet:".length)
-    const { data: pick } = await supabase
-      .from("daily_opportunity_picks").select("id, bet_description, odds").eq("id", pickId).maybeSingle()
-    if (!pick) {
-      await answerCallbackQuery(cq.id, "Essa oportunidade expirou 🕑")
-      return ok("regbet: pick not found")
-    }
-    const unit = await effectiveUnit(supabase, user.id)
-    if (unit) {
-      // COM unidade: atalhos de valor + "Outro valor" (que aí faz sentido)
-      await telegramCall("sendMessage", {
-        chat_id: chatId,
-        text: `📋 Registrar <b>${escHtml(pick.bet_description)}</b> (odd ${Number(pick.odds)})\nQuanto você vai colocar?`,
-        parse_mode: "HTML",
-        reply_markup: { inline_keyboard: [
-          [
-            { text: `1 unidade · ${formatBRL(unit)}`, callback_data: `stk:${pickId}:u1` },
-            { text: `½ unidade · ${formatBRL(unit / 2)}`, callback_data: `stk:${pickId}:uh` }
-          ],
-          [{ text: "Outro valor", callback_data: `stk:${pickId}:ot` }]
-        ] }
-      })
-    } else {
-      // SEM unidade: vai DIRETO pra pergunta (sem botão órfão, sem pergunta dupla)
-      await sendStakePrompt(supabase, chatId, user.id, pickId,
-        `📋 Registrar <b>${escHtml(pick.bet_description)}</b> (odd ${Number(pick.odds)})\nQuanto você colocou? Responde com o valor aqui 👇`)
-    }
-    await answerCallbackQuery(cq.id)
-    return ok("regbet: asked stake")
-  }
-
-  // 💰 stk:<pick_id>:<u1|uh|ot> — escolha do valor
-  if (data.startsWith("stk:")) {
-    const [, pickId, code] = data.split(":")
-    const { data: pick } = await supabase
-      .from("daily_opportunity_picks").select("*").eq("id", pickId).maybeSingle()
-    if (!pick) {
-      await answerCallbackQuery(cq.id, "Essa oportunidade expirou 🕑")
-      return ok("stk: pick not found")
-    }
-
-    // "Outro valor" → pergunta curta por force_reply (o "Registrar X — quanto?"
-    // já foi mostrado com os botões; aqui não repete)
-    if (code === "ot") {
-      await sendStakePrompt(supabase, chatId, user.id, pickId, `Qual valor? Responde com o número aqui 👇`)
-      await answerCallbackQuery(cq.id)
-      return ok("stk: force reply")
-    }
-
-    const unit = await effectiveUnit(supabase, user.id)
-    if (!unit) {
-      await answerCallbackQuery(cq.id, "Configure sua unidade no site ou toque em Outro valor.")
-      return ok("stk: no unit")
-    }
-    const stake = code === "uh" ? Math.round((unit / 2) * 100) / 100 : unit
-    const { data: bet, error } = await registerPickBet(supabase, user.id, pick, stake)
-    if (error || !bet) {
-      await answerCallbackQuery(cq.id, "Deu ruim ao registrar — tenta de novo.")
-      return ok("stk: insert failed")
-    }
-    await editSettledMessage(chatId, messageId, pickReceiptHtml(pick, stake))
-    await answerCallbackQuery(cq.id, "✅ Registrado!")
-    await trackEvent(
-      "opportunity_bet_registered",
-      { bet_id: bet.id, pick_id: pickId, stake, via: "unit_button", channel: "telegram" },
-      user.id, traceId
-    ).catch(() => {})
-    return ok("stk: registered")
-  }
-
-  // ↩️ fix:<bet_id> — desfaz a AUTO-liquidação (notify-settlement): volta pra
-  // pendente e devolve os botões clássicos pro usuário dizer o resultado certo
-  if (data.startsWith("fix:")) {
-    const betId = data.slice("fix:".length)
-
-    const { data: bet } = await supabase
-      .from("bets")
-      .select("id, user_id, status, bet_description, match_description, stake_amount, potential_return, processed_data")
-      .eq("id", betId)
-      .maybeSingle()
-
-    if (!bet || bet.user_id !== user.id) {
-      await answerCallbackQuery(cq.id, "Não achei essa aposta na sua conta.")
-      return ok("fix: bet not found or not owner")
-    }
-    const REVERTIBLE = ["won", "lost", "void", "half_won", "half_lost"] // nunca cashout
-    if (!REVERTIBLE.includes(bet.status)) {
-      await answerCallbackQuery(cq.id, "Essa aposta não está mais liquidada 👍")
-      return ok("fix: invalid status")
-    }
-
-    const pd = bet.processed_data && typeof bet.processed_data === "object" && !Array.isArray(bet.processed_data)
-      ? bet.processed_data
-      : {}
-    const { error: fixErr } = await supabase
-      .from("bets")
-      .update({
-        status: "pending",
-        processed_data: {
-          ...pd,
-          auto_settle: { ...((pd as any).auto_settle ?? {}), corrected_at: new Date().toISOString() },
-        },
-      })
-      .eq("id", betId)
-      .in("status", ["won", "lost", "void", "half_won", "half_lost"]) // nunca reverte cashout
-
-    if (fixErr) {
-      await answerCallbackQuery(cq.id, "Deu ruim ao desfazer — tenta de novo.")
-      return ok("fix: update failed")
-    }
-
-    await telegramCall("editMessageText", {
-      chat_id: chatId,
-      message_id: messageId,
-      text: `↩️ Desfeito — voltou pra pendente.\n\n<b>${escHtml(bet.bet_description)}</b>${bet.match_description ? `\n${escHtml(bet.match_description)}` : ""}\n\nComo foi essa?`,
-      parse_mode: "HTML",
-      disable_web_page_preview: true,
-      reply_markup: {
-        inline_keyboard: [
-          [
-            { text: "✅ Green", callback_data: `settle:${betId}:won` },
-            { text: "❌ Red", callback_data: `settle:${betId}:lost` }
-          ],
-          [{ text: "Outras opções ↗", url: BETS_DASHBOARD_URL }]
-        ]
-      }
-    })
-    await answerCallbackQuery(cq.id, "↩️ Desfeito.")
-    await trackEvent(
-      "auto_settle_corrected",
-      { bet_id: betId, previous_status: bet.status, channel: "telegram" },
-      user.id,
-      traceId
-    ).catch(() => {})
-    return ok("auto settle corrected")
-  }
-
-  // ✅/❌ settle:<bet_id>:<won|lost>
-  if (data.startsWith("settle:")) {
-    const [, betId, outcome] = data.split(":")
-    if (!betId || (outcome !== "won" && outcome !== "lost")) {
-      await answerCallbackQuery(cq.id, "Não entendi esse toque 🤔")
-      return ok("bad callback data")
-    }
-
-    const { data: bet } = await supabase
-      .from("bets")
-      .select("id, user_id, status, bet_description, match_description, stake_amount, potential_return, odds, created_at, channel")
-      .eq("id", betId)
-      .maybeSingle()
-
-    // dono confere? (callback vem do Telegram; a aposta TEM que ser deste usuário)
-    if (!bet || bet.user_id !== user.id) {
-      await answerCallbackQuery(cq.id, "Não achei essa aposta na sua conta.")
-      return ok("bet not found or not owner")
-    }
-
-    if (bet.status !== "pending") {
-      await answerCallbackQuery(cq.id, "Essa já estava registrada 👍")
-      await editSettledMessage(chatId, messageId, settledHtml(bet, bet.status)).catch(() => {})
-      return ok("already settled")
-    }
-
-    const { data: updated, error: updErr } = await supabase
-      .from("bets")
-      .update({ status: outcome })
-      .eq("id", betId)
-      .eq("status", "pending") // guarda otimista: dois toques ≠ duas liquidações
-      .select()
-      .maybeSingle()
-
-    if (updErr || !updated) {
-      await answerCallbackQuery(cq.id, "Deu ruim ao registrar — tenta de novo.")
-      return ok("settle update failed")
-    }
-
-    await editSettledMessage(chatId, messageId, settledHtml(updated, outcome)).catch(() => {})
-    await answerCallbackQuery(cq.id, outcome === "won" ? "✅ Green registrado!" : "❌ Red registrado.")
-    await trackEvent(
-      "settlement_settled_via_bot",
-      { bet_id: betId, outcome, channel: "telegram" },
-      user.id,
-      traceId
-    ).catch(() => {})
-    // Evento canônico da métrica central de retenção (mesmo schema dos 4 caminhos da web em
-    // Bets.tsx). O settlement_settled_via_bot acima fica por continuidade de histórico.
-    await trackEvent(
-      "bet_settled",
-      {
-        product: "betinho",
-        channel: bet.channel ?? "telegram",
-        status: outcome,
-        days_to_settle: bet.created_at
-          ? Math.round((Date.now() - new Date(bet.created_at).getTime()) / 86400000)
-          : null,
-        settled_by: "user_manual",
-        via: "bot_reminder",
-        batch: false,
-        count: 1,
-      },
-      user.id,
-      traceId
-    ).catch(() => {})
-    return ok("bet settled")
-  }
-
-  await answerCallbackQuery(cq.id).catch(() => {})
-  return ok("unknown callback")
-}
-
-async function getDailyBetCount(supabase: any, userId: string, traceId?: string): Promise<number> {
-  try {
-    const nowUTC = new Date()
-    const gmt3Offset = -3 * 60 * 60 * 1000
-    const nowGMT3 = new Date(nowUTC.getTime() + gmt3Offset)
-
-    const year = nowGMT3.getUTCFullYear()
-    const month = String(nowGMT3.getUTCMonth() + 1).padStart(2, "0")
-    const day = String(nowGMT3.getUTCDate()).padStart(2, "0")
-    const gmt3DateString = `${year}-${month}-${day}`
-
-    const startOfDayUTC = new Date(`${gmt3DateString}T03:00:00.000Z`)
-
-    const tomorrowGMT3 = new Date(nowGMT3.getTime() + 24 * 60 * 60 * 1000)
-    const tomorrowYear = tomorrowGMT3.getUTCFullYear()
-    const tomorrowMonth = String(tomorrowGMT3.getUTCMonth() + 1).padStart(2, "0")
-    const tomorrowDay = String(tomorrowGMT3.getUTCDate()).padStart(2, "0")
-    const gmt3TomorrowDateString = `${tomorrowYear}-${tomorrowMonth}-${tomorrowDay}`
-    const startOfTomorrowUTC = new Date(`${gmt3TomorrowDateString}T03:00:00.000Z`)
-
-    console.log("daily_limit_window_calculation", {
-      user_id: userId,
-      trace_id: traceId,
-      now_utc: nowUTC.toISOString(),
-      now_gmt3: nowGMT3.toISOString(),
-      start_of_day_utc: startOfDayUTC.toISOString(),
-      start_of_tomorrow_utc: startOfTomorrowUTC.toISOString(),
-      gmt3_date_string: gmt3DateString
-    })
-
-    if (traceId) {
-      await trackEvent(
-        "daily_limit_window_calculation",
-        {
-          user_id: userId,
-          now_utc: nowUTC.toISOString(),
-          now_gmt3: nowGMT3.toISOString(),
-          start_of_day_utc: startOfDayUTC.toISOString(),
-          start_of_tomorrow_utc: startOfTomorrowUTC.toISOString(),
-          gmt3_date_string: gmt3DateString,
-          channel: "telegram"
-        },
-        userId,
-        traceId
-      ).catch(() => {})
-    }
-
-    const { count, error } = await supabase
-      .from("bets")
-      .select("*", { count: "exact", head: true })
-      .eq("user_id", userId)
-      .gte("bet_date", startOfDayUTC.toISOString())
-      .lt("bet_date", startOfTomorrowUTC.toISOString())
-
-    if (error) {
-      console.error("Error counting daily bets:", error)
-      if (traceId) {
-        await trackEvent(
-          "daily_limit_count_error",
-          {
-            user_id: userId,
-            error: JSON.stringify(error),
-            channel: "telegram"
-          },
-          userId,
-          traceId
-        ).catch(() => {})
-      }
-      return 0
-    }
-
-    const betCount = count || 0
-
-    console.log("daily_limit_bet_count", {
-      user_id: userId,
-      trace_id: traceId,
-      bet_count: betCount,
-      start_of_day_utc: startOfDayUTC.toISOString(),
-      start_of_tomorrow_utc: startOfTomorrowUTC.toISOString()
-    })
-
-    if (traceId) {
-      await trackEvent(
-        "daily_limit_bet_count",
-        {
-          user_id: userId,
-          bet_count: betCount,
-          start_of_day_utc: startOfDayUTC.toISOString(),
-          start_of_tomorrow_utc: startOfTomorrowUTC.toISOString(),
-          channel: "telegram"
-        },
-        userId,
-        traceId
-      ).catch(() => {})
-    }
-
-    return betCount
-  } catch (error) {
-    console.error("Error getting daily bet count:", error)
-    if (traceId) {
-      await trackEvent(
-        "daily_limit_count_exception",
-        {
-          user_id: userId,
-          error_message: error instanceof Error ? error.message : String(error),
-          channel: "telegram"
-        },
-        userId,
-        traceId
-      ).catch(() => {})
-    }
-    return 0
-  }
-}
-
-async function hasReachedDailyLimit(supabase: any, userId: string, traceId?: string): Promise<boolean> {
-  try {
-    console.log("daily_limit_check_start", {
-      user_id: userId,
-      trace_id: traceId,
-      daily_limit: DAILY_BET_LIMIT
-    })
-
-    // Check only betinho subscription status for daily limit
-    const { data: user, error: userError } = await supabase
-      .from("users")
-      .select("betinho_subscription_status")
-      .eq("id", userId)
-      .single()
-
-    if (userError) {
-      console.error("Error fetching user Betinho subscription status:", userError)
-      console.log("daily_limit_user_lookup_error", {
-        user_id: userId,
-        trace_id: traceId,
-        error: JSON.stringify(userError)
-      })
-      if (traceId) {
-        await trackEvent(
-          "daily_limit_user_lookup_error",
-          {
-            user_id: userId,
-            error: JSON.stringify(userError),
-            channel: "telegram"
-          },
-          userId,
-          traceId
-        ).catch(() => {})
-      }
-      return false
-    }
-
-    const subscriptionStatus = user?.betinho_subscription_status || null
-
-    console.log("daily_limit_subscription_status", {
-      user_id: userId,
-      trace_id: traceId,
-      betinho_subscription_status: subscriptionStatus
-    })
-
-    if (traceId) {
-      await trackEvent(
-        "daily_limit_subscription_status",
-        {
-          user_id: userId,
-          betinho_subscription_status: subscriptionStatus,
-          channel: "telegram"
-        },
-        userId,
-        traceId
-      ).catch(() => {})
-    }
-
-    // Premium users (betinho subscription) have no limit
-    if (user?.betinho_subscription_status === "premium") {
-      console.log("daily_limit_premium_bypass", {
-        user_id: userId,
-        trace_id: traceId,
-        betinho_subscription_status: subscriptionStatus
-      })
-      if (traceId) {
-        await trackEvent(
-          "daily_limit_premium_bypass",
-          {
-            user_id: userId,
-            betinho_subscription_status: subscriptionStatus,
-            channel: "telegram"
-          },
-          userId,
-          traceId
-        ).catch(() => {})
-      }
-      return false
-    }
-
-    const betCount = await getDailyBetCount(supabase, userId, traceId)
-    const limitReached = betCount >= DAILY_BET_LIMIT
-
-    console.log("daily_limit_evaluation", {
-      user_id: userId,
-      trace_id: traceId,
-      bet_count: betCount,
-      daily_limit: DAILY_BET_LIMIT,
-      limit_reached: limitReached,
-      betinho_subscription_status: subscriptionStatus
-    })
-
-    if (traceId) {
-      await trackEvent(
-        "daily_limit_evaluation",
-        {
-          user_id: userId,
-          bet_count: betCount,
-          daily_limit: DAILY_BET_LIMIT,
-          limit_reached: limitReached,
-          betinho_subscription_status: subscriptionStatus,
-          channel: "telegram"
-        },
-        userId,
-        traceId
-      ).catch(() => {})
-    }
-
-    return limitReached
-  } catch (error) {
-    console.error("Error checking daily limit:", error)
-    console.log("daily_limit_check_exception", {
-      user_id: userId,
-      trace_id: traceId,
-      error_message: error instanceof Error ? error.message : String(error)
-    })
-    if (traceId) {
-      await trackEvent(
-        "daily_limit_check_exception",
-        {
-          user_id: userId,
-          error_message: error instanceof Error ? error.message : String(error),
-          channel: "telegram"
-        },
-        userId,
-        traceId
-      ).catch(() => {})
-    }
-    return false
-  }
-}
-
-async function transcribeAudio(audioUrl: string, userId: string, traceId: string): Promise<string> {
-  const startTime = Date.now()
-
-  try {
-    const response = await fetch(`${OPENAI_API_URL}/audio/transcriptions`, {
-      method: "POST",
-      headers: {
-        Authorization: `Bearer ${OPENAI_API_KEY}`,
-        "Content-Type": "application/json"
-      },
-      body: JSON.stringify({
-        file: audioUrl,
-        model: "whisper-1",
-        language: "pt"
-      })
-    })
-
-    const latency = Date.now() - startTime
-
-    if (!response.ok) {
-      const errorText = await response.text()
-      await trackEvent(
-        "openai_api_error",
-        { error_type: "transcribe_audio_error", error_message: errorText.substring(0, 500), model: "whisper-1", operation: "transcribe_audio", channel: "telegram" },
-        userId,
-        traceId
-      ).catch(() => {})
-      throw new Error(`OpenAI API error: ${response.status} - ${errorText}`)
-    }
-
-    const result = await response.json()
-    const estimatedUsage = { prompt_tokens: 0, completion_tokens: 0, total_tokens: 0 }
-    await trackLLMGeneration(
-      "transcribe_audio",
-      "whisper-1",
-      [{ role: "user", content: audioUrl }],
-      { choices: [{ message: { content: result.text } }] },
-      estimatedUsage,
-      latency,
-      userId,
-      traceId,
-      { channel: "telegram" }
-    ).catch(() => {})
-
-    return result.text
-  } catch (error) {
-    await trackEvent(
-      "processing_error",
-      { error_type: "transcribe_audio_error", error_message: error.message?.substring(0, 500), operation: "transcribe_audio", channel: "telegram" },
-      userId,
-      traceId
-    ).catch(() => {})
-    throw new Error(`Failed to transcribe audio: ${error.message}`)
-  }
-}
-
-async function processImage(imageUrl: string, userId: string, traceId: string): Promise<string> {
-  const startTime = Date.now()
-
-  try {
-    const messages = [{
-      role: "user" as const,
-      content: [
-        {
-          type: "text",
-          text: `Analise esta imagem de aposta esportiva e extraia TODAS as informações possíveis em português.
-
-CRÍTICO - Reconhecer Múltiplas Seções de Apostas Múltiplas:
-- Se a imagem mostrar VÁRIAS seções de "CRIAR APOSTA" ou múltiplas apostas COM SUAS PRÓPRIAS ODDS COMBINADAS (ex: uma seção mostra "1.86" e outra mostra "4.60"), cada seção é uma MÚLTIPLA APOSTA separada com sua odd já combinada
-- Cada seção de "CRIAR APOSTA" com uma odd mostrada (como "1.86" ou "4.60") representa UMA múltipla aposta já combinada
-- Se houver múltiplas seções, você precisa identificar a ODD COMBINADA de cada seção (não as odds individuais dos jogadores)
-
-CASO 1: Uma única seção com múltiplas apostas individuais
-- Se houver apenas UMA seção com vários jogadores e uma odd combinada no topo (ex: "3.20")
-- Extraia cada jogador individualmente com suas odds individuais
-- Exemplo: "Jogador 1 - Tipo - Odd: 2.80, Jogador 2 - Tipo - Odd: 3.20"
-
-CASO 2: Múltiplas seções de "CRIAR APOSTA" com odds combinadas
-- Se houver VÁRIAS seções de "CRIAR APOSTA", cada uma com sua própria odd combinada
-- Para cada seção, identifique a ODD COMBINADA mostrada (ex: "1.86" ou "4.60")
-- Liste cada seção como uma "perna" com a odd combinada
-- Exemplo: "SEÇÃO 1: Odd combinada 1.86 (Royce O'Neale + Devin Booker)"
-         "SEÇÃO 2: Odd combinada 4.60 (Jerami Grant + Deni Avdija)"
-
-Para cada seção/aposta identifique:
-- Times/atletas envolvidos
-- Tipo de aposta (pontos, assistências, rebotes, etc.)
-- Se houver odd combinada mostrada claramente na seção, use essa odd (é uma múltipla aposta já calculada)
-- Se não houver odd combinada, extraia odds individuais
-- Valor apostado (se visível, geralmente no final)
-- Esporte e liga/campeonato
-
-Retorne apenas o texto estruturado, sem explicações adicionais.`
-        },
-        { type: "image_url", image_url: { url: imageUrl } }
-      ]
-    }]
-
-    const response = await fetch(`${OPENAI_API_URL}/chat/completions`, {
-      method: "POST",
-      headers: { Authorization: `Bearer ${OPENAI_API_KEY}`, "Content-Type": "application/json" },
-      body: JSON.stringify({ model: "gpt-4o", messages, max_tokens: 800 })
-    })
-
-    const latency = Date.now() - startTime
-
-    if (!response.ok) {
-      const errorText = await response.text()
-      await trackEvent(
-        "openai_api_error",
-        { error_type: "process_image_error", error_message: errorText.substring(0, 500), model: "gpt-4o", operation: "process_image", channel: "telegram" },
-        userId,
-        traceId
-      ).catch(() => {})
-      throw new Error(`OpenAI API error: ${response.status} - ${errorText}`)
-    }
-
-    const result = await response.json()
-    const usage = result.usage || { prompt_tokens: 0, completion_tokens: 0, total_tokens: 0 }
-    await trackLLMGeneration(
-      "process_image",
-      "gpt-4o",
-      messages,
-      result,
-      usage,
-      latency,
-      userId,
-      traceId,
-      { channel: "telegram" }
-    ).catch(() => {})
-
-    return result.choices[0].message.content
-  } catch (error) {
-    await trackEvent(
-      "processing_error",
-      { error_type: "process_image_error", error_message: error.message?.substring(0, 500), operation: "process_image", channel: "telegram" },
-      userId,
-      traceId
-    ).catch(() => {})
-    throw new Error(`Failed to process image: ${error.message}`)
-  }
-}
-
-async function extractBettingInfo(content: string, userId: string, traceId: string): Promise<ProcessedBet | null> {
-  try {
-    console.log("extract_betting_request", { content_preview: content.slice(0, 200) })
-    const prompt = `Analise a seguinte mensagem e determine se é uma aposta esportiva válida. Se NÃO for uma aposta, retorne null imediatamente.
-
-MENSAGEM: "${content}"
-
-IMPORTANTE - VALIDAÇÃO PRÉVIA:
-- Se a mensagem for apenas um cumprimento (ex: "oi", "olá", "tudo bem"), retorne null
-- Se a mensagem não mencionar nenhum time/jogador/esporte, retorne null
-- Se a mensagem não mencionar odds ou valores relacionados a apostas, retorne null
-- Se a mensagem for muito curta e não contiver informações de aposta, retorne null
-- APENAS extraia informações se for claramente uma aposta esportiva com odds, times/jogadores e tipo de aposta
-
-INSTRUÇÕES (APENAS SE FOR UMA APOSTA VÁLIDA):
-1. Identifique se é uma aposta simples, múltipla ou sistema
-   - "single" = 1 única aposta
-   - "multiple" = 2 ou mais apostas combinadas
-   - "system" = sistema de apostas
-2. Extraia TODOS os jogos/eventos mencionados
-   - Cada JOGO diferente = uma entrada no array matches
-   - Se múltiplos jogadores estão no MESMO JOGO e há uma "ODD COMBINADA" no topo, combine-os em UMA única entrada
-   - Se jogadores estão em jogos diferentes, cada jogo = uma entrada separada
-3. Para cada match, identifique: times/jogo, atletas envolvidos (se múltiplos no mesmo jogo), tipo de aposta, e a odd
-4. Identifique o valor apostado (stake)
-5. Identifique datas quando mencionadas
-
-IDENTIFICAÇÃO DE TIPO DE ODDS (CRÍTICO):
-
-CENÁRIO A: Múltiplas seções com odds já combinadas
-- Se a mensagem mencionar MÚLTIPLAS "SEÇÃO" ou "CRIAR APOSTA" com odds diferentes (ex: "SEÇÃO 1: Odd 1.86", "SEÇÃO 2: Odd 4.60")
-- Cada seção é uma perna com sua odd já combinada
-- Marque "odds_are_individual": false
-- Marque "is_combined_odd": true para cada match que tiver odd combinada
-
-CENÁRIO B: Uma única seção com múltiplas apostas individuais
-- Se houver múltiplos jogadores/atletas listados, mas cada um tem sua própria odd individual claramente mostrada
-- Cada jogador/atleta deve ser uma entrada separada no array matches
-- Marque "odds_are_individual": true
-- Marque "is_combined_odd": false
-
-CENÁRIO C: Apenas uma aposta simples
-- Marque "bet_type": "single"
-- Marque "odds_are_individual": true
-- "is_combined_odd": false
-
-REQUISITOS DE SAÍDA:
-- Sem texto extra, retorne JSON válido seguindo o schema abaixo
-- matches é um array de jogos; odds sempre >= 1.01
-- stake_amount pode ser 0 se não informado
-
-CLASSIFICAÇÃO DE ESPORTE:
- - Identifique o esporte com base nas informações da aposta
- - Padronize e retorne o nome do esporte de acordo com a lista abaixo:
-  * Futebol
-  * Basquete
-  * Tênis
-  * Futebol Americano
-  * Futsal
-  * Vôlei
-  * eSports
-  * MMA/UFC
-  * Boxe
-  * Hóquei no Gelo
-  * Beisebol
-  * Golfe
-  * Tênis de Mesa
-  * Handebol
-  * Rugby
-  * Corrida de Cavalos
-  * Ciclismo
-  * Críquete
-  * Dardos
-  * Snooker
-  * Badminton
-  * Futebol Australiano
-  * Esqui
-  * Biatlo
-  * Automobilismo
-  * Vôlei de Praia
-  * Padel
-  * Natação
-  * Atletismo
- - Se não for possível identificar o esporte, retorne o esporte como string vazia.
- - Se o esporte for identificado e não estiver na lista, retorne ele como está.
-
-CLASSIFICAÇÃO DE LIGAS:
- - Identifique a liga com base nas informações da aposta
- - Padronize e retorne o nome da liga de acordo com o campo "league" do seguinte json:
-
-[
-  {
-    "league": "US - NBA",
-    "sport": "Futebol"
-  },
-  {
-    "league": "BR - Copa do Brasil",
-    "sport": "Futebol"
-  },
-  {
-    "league": "US - NBA",
-    "sport": "Basquete"
-  },
-  {
-    "league": "EU - Europa League",
-    "sport": "Futebol"
-  },
-  {
-    "league": "Mundial de Clubes FIFA",
-    "sport": "Futebol"
-  },
-  {
-    "league": "EN - Premier League",
-    "sport": "Futebol"
-  },
-  {
-    "league": "SAU - Pro League",
-    "sport": "Futebol"
-  },
-  {
-    "league": "AU - NBL",
-    "sport": "Basquete"
-  },
-  {
-    "league": "ITA - Série A",
-    "sport": "Futebol"
-  },
-  {
-    "league": "EU - Champions League",
-    "sport": "Futebol"
-  },
-  {
-    "league": "US - NFL",
-    "sport": "Futebol Americano"
-  },
-  {
-    "league": "ME - Liga Premier",
-    "sport": "Futebol"
-  },
-  {
-    "league": "EU - Conference League",
-    "sport": "Futebol"
-  },
-  {
-    "league": "AME - Copa Sul-Americana",
-    "sport": "Futebol"
-  },
-  {
-    "league": "AME - Copa Libertadores",
-    "sport": "Futebol"
-  },
-  {
-    "league": "AL - Bundesliga",
-    "sport": "Futebol"
-  },
-  {
-    "league": "FR - Ligue 1",
-    "sport": "Futebol"
-  },
-  {
-    "league": "Fórmula 1",
-    "sport": "Automobilismo"
-  },
-  {
-    "league": "HOL - Eerste Divisie",
-    "sport": "Futebol"
-  },
-  {
-    "league": "BR - Série B",
-    "sport": "Futebol"
-  },
-  {
-    "league": "TUR - Lig 1",
-    "sport": "Futebol"
-  },
-  {
-    "league": "ES - La Liga",
-    "sport": "Futebol"
-  },
-  {
-    "league": "EU - Eliminatória UEFA",
-    "sport": "Futebol"
-  },
-  {
-    "league": "PT - Primeira Liga",
-    "sport": "Futebol"
-  },
-  {
-    "league": "BR - Série A",
-    "sport": "Futebol"
-  },
-  {
-    "league": "BR - Paulistão",
-    "sport": "Futebol"
-  },
-  {
-    "league": "BEL - Pro League",
-    "sport": "Futebol"
-  }
-]
-- Se não for possível identificar a liga, retorne a liga como string vazia.
-- Se a liga for identificada e não estiver na lista, retorne ela como está.
-- Se forem identificadas múltiplas ligas, retorne "Diversos"
-
-CLASSIFICAÇÃO DE MERCADO DA APOSTA:
-- Identifique o mercado da aposta com base na descrição (tipo de aposta, odds, termos usados).
-- Valores permitidos (use exatamente um):
-  * "Múltipla" = múltiplas seleções combinadas (acumulada)
-  * "Money Line" = aposta no vencedor do jogo/evento (1x2 em futebol, winner)
-  * "Handicap" = linha de handicap (asiático ou europeu)
-  * "Over/Under" = totais (over/under gols, pontos)
-  * "Dupla Chance" = duas opções (ex: 1X, X2, 12)
-  * "Ambas Marcam" = sim/não (ambas marcam, BTTS)
-- Retorne string vazia "" se não for possível identificar o mercado.
-
-SCHEMA:
-{
-  bet_type: "single" | "multiple" | "system",
-  sport: string,
-  league: string | null,
-  betting_market: string,
-  matches: [{
-    description: string,
-    bet_description: string,
-    odds: number,
-    match_date: string | null,
-    is_combined_odd: boolean
-  }],
-  stake_amount: number,
-  bet_date: string,
-  odds_are_individual: boolean
-}`
-
-    const messages = [{ role: "user" as const, content: prompt }]
-
-    const startTime = Date.now()
-
-    const response = await fetch(`${OPENAI_API_URL}/chat/completions`, {
-      method: "POST",
-      headers: { Authorization: `Bearer ${OPENAI_API_KEY}`, "Content-Type": "application/json" },
-      body: JSON.stringify({
-        model: "gpt-4o",
-        messages,
-        tools: [CALCULATE_ODDS_TOOL],
-        tool_choice: "auto",
-        response_format: {
-          type: "json_schema",
-          json_schema: { name: "betting_info", strict: true, schema: BETTING_INFO_SCHEMA }
-        },
-        max_tokens: 2000,
-        temperature: 0.1
-      })
-    })
-
-    const initialLatency = Date.now() - startTime
-
-    if (!response.ok) {
-      const errorText = await response.text()
-      await trackEvent(
-        "openai_api_error",
-        { error_type: "extract_betting_error", error_message: errorText.substring(0, 500), model: "gpt-4o", operation: "extract_betting", channel: "telegram" },
-        userId,
-        traceId
-      ).catch(() => {})
-      throw new Error(`OpenAI API error: ${response.status} - ${errorText}`)
-    }
-
-    let result = await response.json()
-    let message = result.choices[0].message
-    let totalLatency = initialLatency
-    let totalUsage = result.usage || { prompt_tokens: 0, completion_tokens: 0, total_tokens: 0 }
-    const usedToolCalling = !!(message.tool_calls && message.tool_calls.length > 0)
-
-    console.log("extract_betting_first_choice", {
-      content_preview: typeof message.content === "string" ? message.content.slice(0, 200) : "",
-      tool_calls_len: message.tool_calls?.length || 0
-    })
-
-    if (message.tool_calls && message.tool_calls.length > 0) {
-      try {
-        const toolResponses: Array<{ role: "tool"; tool_call_id: string; content: string }> = []
-        for (const toolCall of message.tool_calls) {
-          const id = toolCall.id
-          const name = toolCall.function?.name
-          const argsJson = toolCall.function?.arguments || "{}"
-          if (name === "calculate_multiple_odds") {
-            const args = JSON.parse(argsJson)
-            const odds = (args?.odds as number[]) ?? []
-            const combinedOdd = await calculateMultipleOdds(odds)
-            toolResponses.push({ role: "tool" as const, tool_call_id: id, content: JSON.stringify({ combined_odd: combinedOdd }) })
-          } else {
-            toolResponses.push({ role: "tool" as const, tool_call_id: id, content: JSON.stringify({ error: "unknown_tool" }) })
-          }
-        }
-
-        const followUp = await fetch(`${OPENAI_API_URL}/chat/completions`, {
-          method: "POST",
-          headers: { Authorization: `Bearer ${OPENAI_API_KEY}`, "Content-Type": "application/json" },
-          body: JSON.stringify({
-            model: "gpt-4o",
-            messages: [...messages, message, ...toolResponses],
-            tools: [CALCULATE_ODDS_TOOL],
-            tool_choice: "none",
-            response_format: { type: "json_schema", json_schema: { name: "betting_info", strict: true, schema: BETTING_INFO_SCHEMA } },
-            max_tokens: 1500,
-            temperature: 0.1
-          })
-        })
-
-        const followLatency = Date.now() - startTime - initialLatency
-        const followResult = await followUp.json()
-        totalLatency += followLatency
-        const followUsage = followResult.usage || { prompt_tokens: 0, completion_tokens: 0, total_tokens: 0 }
-        totalUsage = {
-          prompt_tokens: (totalUsage.prompt_tokens || 0) + (followUsage.prompt_tokens || 0),
-          completion_tokens: (totalUsage.completion_tokens || 0) + (followUsage.completion_tokens || 0),
-          total_tokens: (totalUsage.total_tokens || 0) + (followUsage.total_tokens || 0)
-        }
-
-        const followChoice = followResult?.choices?.[0]
-        if (!followChoice?.message) {
-          const apiError = followResult?.error?.message || JSON.stringify(followResult).slice(0, 500)
-          console.error("extract_betting_follow_no_choices", { followResult: followResult?.error || followResult })
-          await trackEvent(
-            "processing_error",
-            { error_type: "tool_call_error", error_message: apiError, operation: "extract_betting", channel: "telegram" },
-            userId,
-            traceId
-          ).catch(() => {})
-          return null
-        }
-
-        message = followChoice.message
-        result = followResult
-        console.log("extract_betting_after_tool", {
-          content_preview: typeof message.content === "string" ? message.content.slice(0, 200) : "",
-          tool_calls_len: message.tool_calls?.length || 0
-        })
-      } catch (err) {
-        await trackEvent(
-          "processing_error",
-          { error_type: "tool_call_error", error_message: (err as any)?.message?.substring(0, 500), operation: "extract_betting", channel: "telegram" },
-          userId,
-          traceId
-        ).catch(() => {})
-        return null
-      }
-    }
-
-    await trackLLMGeneration(
-      "extract_betting",
-      "gpt-4o",
-      messages,
-      result,
-      totalUsage,
-      totalLatency,
-      userId,
-      traceId,
-      { channel: "telegram", used_tool_calling: usedToolCalling }
-    ).catch(() => {})
-
-    console.log("extract_betting_raw_content", {
-      content_preview: typeof message.content === "string" ? message.content.slice(0, 300) : "",
-      tool_calls: message.tool_calls?.length || 0
-    })
-
-    let bettingInfo: ProcessedBet
-    try {
-      bettingInfo = JSON.parse(message.content) as ProcessedBet
-    } catch (err) {
-      console.error("extract_betting_parse_error", { error: (err as any)?.message, content: message.content })
-      await trackEvent(
-        "processing_error",
-        { error_type: "extract_betting_parse_error", error_message: (err as any)?.message?.substring(0, 500), operation: "extract_betting", channel: "telegram" },
-        userId,
-        traceId
-      ).catch(() => {})
-      return null
-    }
-    if (!bettingInfo.matches || bettingInfo.matches.length === 0) {
-      console.log("extract_betting_no_matches", {
-        content_preview: message.content?.slice ? message.content.slice(0, 200) : "",
-        raw_message_type: typeof message.content
-      })
-      await trackEvent(
-        "bet_extraction_failed",
-        { message: "no_matches_extracted", channel: "telegram" },
-        userId,
-        traceId
-      ).catch(() => {})
-      return null
-    }
-    return bettingInfo
-  } catch (error) {
-    await trackEvent(
-      "processing_error",
-      { error_type: "extract_betting_error", error_message: (error as any)?.message?.substring(0, 500), operation: "extract_betting", channel: "telegram" },
-      userId,
-      traceId
-    ).catch(() => {})
-    return null
-  }
-}
-
-async function processMessage(
-  supabase: any,
-  messageId: string,
-  content: string,
-  userId: string,
-  traceId: string,
-  chatId: string | number
-): Promise<void> {
-  try {
-    await supabase.from("message_queue").update({ status: "processing" }).eq("id", messageId)
-
-    const normalizedContent = content.toLowerCase().trim()
-    const simpleGreetings = ["oi", "olá", "ola", "hello", "hi", "hey", "eae", "e aí", "eai", "tudo bem", "tudo bom", "ok", "okay", "beleza", "blz"]
-    const isSimpleGreeting = simpleGreetings.some(greeting => normalizedContent === greeting)
-    const isTooShort = normalizedContent.length < 10 && !normalizedContent.match(/\d/)
-    const bettingKeywords = ["odd", "odds", "aposta", "apost", "jogo", "time", "times", "vs", "pontos", "gols", "assist", "rebote", "stake", "valor", "reais", "r$", "bet", "betting"]
-    const hasBettingKeywords = bettingKeywords.some(keyword => normalizedContent.includes(keyword))
-
-    console.log("process_message_start", {
-      message_id: messageId,
-      user_id: userId,
-      content_length: content.length,
-      isSimpleGreeting,
-      isTooShort,
-      hasBettingKeywords
-    })
-
-    if (isSimpleGreeting || (isTooShort && !hasBettingKeywords)) {
-      await trackEvent(
-        "bet_extraction_skipped",
-        { message_id: messageId, reason: isSimpleGreeting ? "simple_greeting" : "too_short_without_keywords", content_length: content.length, channel: "telegram" },
-        userId,
-        traceId
-      ).catch(() => {})
-
-      await supabase.from("message_queue").update({ status: "failed", error_message: "Message does not appear to be a betting message" }).eq("id", messageId)
-      await sendHelpMessageTelegram(chatId)
-      return
-    }
-
-    await trackEvent(
-      "bet_extraction_started",
-      { message_id: messageId, content_length: content.length, channel: "telegram" },
-      userId,
-      traceId
-    ).catch(() => {})
-
-    let bettingInfo = await extractBettingInfo(content, userId, traceId)
-    console.log("bet_extraction_result", {
-      message_id: messageId,
-      has_betting_info: !!bettingInfo,
-      matches_count: bettingInfo?.matches?.length,
-      bet_type: bettingInfo?.bet_type
-    })
-
-    if (bettingInfo && bettingInfo.matches && bettingInfo.matches.length > 0) {
-      const validMatches = bettingInfo.matches.filter(match =>
-        match.description &&
-        match.description.trim().length > 0 &&
-        match.bet_description &&
-        match.bet_description.trim().length > 0 &&
-        match.odds &&
-        match.odds >= 1.01
-      )
-
-      console.log("bet_validation_matches", {
-        message_id: messageId,
-        original_count: bettingInfo.matches.length,
-        valid_count: validMatches.length
-      })
-
-      if (validMatches.length === 0) {
-        console.log("bet_extraction_valid_matches_empty", { message_id: messageId, raw_matches: bettingInfo.matches })
-        await trackEvent(
-          "bet_extraction_failed",
-          { message_id: messageId, reason: "invalid_matches_extracted", channel: "telegram" },
-          userId,
-          traceId
-        ).catch(() => {})
-
-        await supabase.from("message_queue").update({ status: "failed", error_message: "Extracted matches are invalid" }).eq("id", messageId)
-        await sendHelpMessageTelegram(chatId)
-        return
-      }
-
-      const limitReached = await hasReachedDailyLimit(supabase, userId, traceId)
-
-      console.log("daily_limit_check_result", {
-        user_id: userId,
-        message_id: messageId,
-        trace_id: traceId,
-        limit_reached: limitReached,
-        daily_limit: DAILY_BET_LIMIT
-      })
-
-      if (limitReached) {
-        await trackEvent(
-          "daily_limit_reached",
-          { user_id: userId, daily_limit: DAILY_BET_LIMIT, channel: "telegram" },
-          userId,
-          traceId
-        ).catch(() => {})
-
-        await supabase
-          .from("message_queue")
-          .update({ status: "completed", processed_at: new Date().toISOString(), error_message: "Daily bet limit reached" })
-          .eq("id", messageId)
-
-        await sendPaywallMessageTelegram(chatId)
-        return
-      }
-
-      const MAX_INDIVIDUAL_ODD = 20.0
-
-      // BUGFIX: Use validMatches instead of bettingInfo.matches for bet_type check
-      if (validMatches.length > 1 && bettingInfo.bet_type !== "multiple") {
-        bettingInfo.bet_type = "multiple"
-      }
-
-      // BUGFIX: Use validMatches instead of bettingInfo.matches for validation
-      const validatedMatches = validMatches.map((match) => {
-        const odd = match.odds || 1
-        if (odd > MAX_INDIVIDUAL_ODD) {
-          trackEvent(
-            "bet_validation_warning",
-            { warning_type: "high_individual_odd", odds_value: odd, channel: "telegram" },
-            userId,
-            traceId
-          ).catch(() => {})
-          return { ...match, odds: Math.min(odd, MAX_INDIVIDUAL_ODD), original_odd: odd }
-        }
-        if (odd < 1.01) {
-          trackEvent(
-            "bet_validation_warning",
-            { warning_type: "low_odd", odds_value: odd, channel: "telegram" },
-            userId,
-            traceId
-          ).catch(() => {})
-          return { ...match, odds: 1.01 }
-        }
-        return match
-      })
-
-      let totalOdds = 1
-
-      if (bettingInfo.bet_type === "multiple" && validatedMatches.length > 1) {
-        totalOdds = validatedMatches.reduce((acc, match) => acc * (match.odds || 1), 1)
-      } else if (validatedMatches.length === 1) {
-        totalOdds = validatedMatches[0]?.odds || 1
-      } else {
-        totalOdds = validatedMatches.reduce((acc, match) => acc * (match.odds || 1), 1)
-      }
-
-      const currentDate = new Date().toISOString()
-      const stakeAmount = bettingInfo.stake_amount || 0
-      // BUGFIX: Use validatedMatches for fallback instead of bettingInfo.matches
-      const calculatedOdds = bettingInfo.bet_type === "multiple" ? totalOdds : (validatedMatches[0]?.odds || 1.01)
-
-      bettingInfo.matches = validatedMatches
-
-      console.log("bet_insert_starting", {
-        message_id: messageId,
-        user_id: userId,
-        bet_type: bettingInfo.bet_type,
-        stake_amount: stakeAmount,
-        calculated_odds: calculatedOdds,
-        matches_count: validatedMatches.length
-      })
-
-      const { data: bet, error: betError } = await supabase
-        .from("bets")
-        .insert({
-          user_id: userId,
-          bet_type: bettingInfo.bet_type,
-          sport: bettingInfo.sport || "outros",
-          league: bettingInfo.league || null,
-          betting_market: bettingInfo.betting_market || null,
-          match_description: validatedMatches.length === 1 ? validatedMatches[0]?.description : `Múltipla (${validatedMatches.length} seleções)`,
-          bet_description: validatedMatches.length === 1 ? validatedMatches[0]?.bet_description : validatedMatches.map((m, i) => `${m.description} - ${m.bet_description}`).join(" • "),
-          odds: calculatedOdds,
-          stake_amount: stakeAmount,
-          potential_return: stakeAmount * calculatedOdds,
-          bet_date: currentDate,
-          match_date: validatedMatches[0]?.match_date || null,
-          raw_input: content,
-          processed_data: bettingInfo,
-          channel: "telegram"
-        })
-        .select()
-        .single()
-
-      if (betError) {
-        console.error("bet_insert_error", { message_id: messageId, error: betError })
-        throw betError
-      }
-
-      console.log("bet_insert_success", { message_id: messageId, bet_id: bet.id })
-
-      await trackEvent(
-        "bet_created",
-        { bet_id: bet.id, bet_type: bettingInfo.bet_type, total_odds: calculatedOdds, stake_amount: stakeAmount, legs_count: validatedMatches.length, channel: "telegram" },
-        userId,
-        traceId
-      ).catch(() => {})
-
-      if (bettingInfo.bet_type === "multiple" && validatedMatches.length > 1) {
-        for (let i = 0; i < validatedMatches.length; i++) {
-          const match = validatedMatches[i]
-          if (!match.description || !match.bet_description || !match.odds) continue
-          await supabase
-            .from("bet_legs")
-            .insert({
-              bet_id: bet.id,
-              leg_number: i + 1,
-              sport: bettingInfo.sport || "outros",
-              match_description: match.description,
-              bet_description: match.bet_description,
-              odds: match.odds,
-              status: "pending"
-            })
-        }
-      }
-
-      await supabase
-        .from("message_queue")
-        .update({ status: "completed", processed_at: new Date().toISOString() })
-        .eq("id", messageId)
-
-      console.log("bet_sending_confirmation", { message_id: messageId, bet_id: bet.id, chat_id: chatId })
-      await sendConfirmationMessageTelegram(chatId, bet)
-      console.log("bet_confirmation_sent", { message_id: messageId, bet_id: bet.id })
-    } else {
-      await trackEvent(
-        "bet_extraction_failed",
-        { message_id: messageId, reason: "no_matches_extracted", channel: "telegram" },
-        userId,
-        traceId
-      ).catch(() => {})
-
-      await supabase.from("message_queue").update({ status: "failed", error_message: "Could not extract betting information from message" }).eq("id", messageId)
-      await sendHelpMessageTelegram(chatId)
-    }
-  } catch (error) {
-    await trackEvent(
-      "processing_error",
-      { error_type: "message_processing_error", error_message: error.message?.substring(0, 500), message_id: messageId, channel: "telegram" },
-      userId,
-      traceId
-    ).catch(() => {})
-
-    await supabase.from("message_queue").update({ status: "failed", error_message: error.message }).eq("id", messageId)
-  }
-}
-
-async function findUserByTelegram(
-  supabase: any,
-  telegramUserId?: string,
-  chatId?: string,
-  traceId?: string
-): Promise<{ id: string; name?: string | null; telegram_chat_id?: string | null; whatsapp_number?: string | null } | null> {
-  if (!telegramUserId && !chatId) {
-    console.log("daily_limit_user_resolution_missing_ids", {
-      telegram_user_id: telegramUserId,
-      chat_id: chatId,
-      trace_id: traceId
-    })
-    return null
-  }
-
-  const ors: string[] = []
-  if (chatId) ors.push(`telegram_chat_id.eq.${chatId}`)
-  if (telegramUserId) ors.push(`telegram_user_id.eq.${telegramUserId}`)
-
-  if (ors.length === 0) {
-    console.log("daily_limit_user_resolution_no_conditions", {
-      telegram_user_id: telegramUserId,
-      chat_id: chatId,
-      trace_id: traceId
-    })
-    return null
-  }
-
-  console.log("daily_limit_user_resolution_start", {
-    telegram_user_id: telegramUserId,
-    chat_id: chatId,
-    trace_id: traceId,
-    query_conditions: ors.join(",")
-  })
-
-  const { data, error } = await supabase
-    .from("users")
-    .select("id, name, telegram_chat_id, whatsapp_number")
-    .or(ors.join(","))
-    .limit(1)
-    .maybeSingle()
-
-  if (error) {
-    console.log("daily_limit_user_resolution_error", {
-      telegram_user_id: telegramUserId,
-      chat_id: chatId,
-      trace_id: traceId,
-      error: JSON.stringify(error)
-    })
-    return null
-  }
-
-  if (!data) {
-    console.log("daily_limit_user_resolution_not_found", {
-      telegram_user_id: telegramUserId,
-      chat_id: chatId,
-      trace_id: traceId
-    })
-    return null
-  }
-
-  console.log("daily_limit_user_resolution_success", {
-    telegram_user_id: telegramUserId,
-    chat_id: chatId,
-    resolved_user_id: data.id,
-    trace_id: traceId
-  })
-
-  return data
-}
-
-async function findUserByPhone(supabase: any, phone: string): Promise<any | null> {
-  if (!phone) return null
-  const normalizedCandidates = normalizePhoneCandidates(phone)
-  const { data: users } = await supabase
-    .from("users")
-    .select("id, whatsapp_number, telegram_phone, name")
-
-  if (!users) return null
-
-  const found = users.find((u: any) => {
-    const userPhones = [u.whatsapp_number, u.telegram_phone].filter(Boolean) as string[]
-    const match = userPhones.some(p => {
-      const userVariants = normalizePhoneCandidates(p)
-      return userVariants.some(uv => normalizedCandidates.includes(uv))
-    })
-    if (match) {
-      console.log('phone_match_found', {
-        incoming_phone: maskPhone(phone),
-        incoming_candidates: normalizedCandidates,
-        user_id: u.id,
-        user_phones: userPhones.map(maskPhone)
-      })
-    } else {
-      console.log('phone_match_miss', {
-        incoming_phone: maskPhone(phone),
-        incoming_candidates: normalizedCandidates,
-        user_id: u.id,
-        user_phones: userPhones.map(maskPhone)
-      })
-    }
-    return match
-  }) || null
-
-  if (!found) {
-    console.log('phone_match_none', {
-      incoming_phone: maskPhone(phone),
-      incoming_candidates: normalizedCandidates,
-      users_checked: users.length
-    })
-  }
-
-  return found
-}
 
 serve(async (req) => {
   if (req.method === "OPTIONS") {
@@ -2424,4 +532,5 @@ serve(async (req) => {
     return new Response(JSON.stringify({ success: false, error: error.message, timestamp: new Date().toISOString() }), { headers: { "Content-Type": "application/json" }, status: 500 })
   }
 })
+
